+++ public/changelog.md
@@ -15,7 +15,7 @@
 - **add show revealed cards to opponent**
 - **fix Discord bot lobby creation from queue**
 - **save camera and playmat settings**
-- **fix import list to collection**
+- **fix import list to collection** and make card display nicer
 
 ## January 2026